// src/app/shared/interfaces/game-events.interface.ts
export interface GameEvents {
  homeGoals: any[];
  awayGoals: any[];
  homeYellowCards: any[];
  awayYellowCards: any[];
  homeSecondYellowCards: any[];
  awaySecondYellowCards: any[];
  homeRedCards: any[];
  awayRedCards: any[];
  allGoals: any[];
  allYellowCards: any[];
  allSecondYellowCards: any[];
  allRedCards: any[];
  allEvents: any[];
  firstHalfEvents: any[];
  secondHalfEvents: any[];
}